requirements.txt
# =====================================
# Series Architect - Python Dependencies
# =====================================

# Data Layer Dependencies
psycopg2-binary>=2.9.0
requests>=2.31.0
python-dotenv>=1.0.0

# Logic Layer Dependencies
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Optional: Development Tools
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0

streamlit~=1.54.0
flask~=3.1.2